Models(added properties to classes to allow better parsing),Db.cs(added basic Put,Get functions wtih a helper PrepareCommands

+++ Technical1/FileOps/IO.cs
@@ -98,7 +98,7 @@ namespace Technical1.FileOps
             string Current_Date = DateTime.Today.ToString("MM/dd/yyyy");
 
             int invoiceRecordCount = billList.Count;
-            double invoiceRecordTotalAmount = default;
+            decimal invoiceRecordTotalAmount = default;
 
             foreach (BillHeader b in billList)
             {
